requirements.txt
acme==2.9.0
appdirs==1.4.4
attrs==23.2.0
autocommand==2.2.2
Automat==22.10.0
authlib==1.5.2
Babel==2.14.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.46
botocore==1.34.46
Brotli==1.1.0
certbot==2.9.0
certbot-nginx==2.9.0
certifi==2024.6.2
chardet==5.2.0
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
configobj==5.0.8
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docker==7.1.0
dockerpty==0.4.1
docopt==0.6.2
fastapi>=0.70.0
fonttools==4.46.0
fs==2.4.16
Glances==4.0.5
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
inflect==7.3.1
influxdb==5.3.2
jaraco.context==6.0.0
jaraco.functools==4.0.2
Jinja2==3.1.3
jmespath==1.0.1
josepy==1.14.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema>=2.5.1,<4
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
launchpadlib==2.0.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.3
netaddr==0.10.1
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
passlib[bcrypt]
python-multipart
packaging==24.1
parsedatetime==2.6
pexpect==4.9.0
pillow==10.4.0
ply==3.11
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyasyncore==1.0.2
pycryptodomex==3.20.0
Pygments==2.18.0
PyGObject==3.48.2
PyHamcrest==2.1.0
PyICU==2.13.1
pyinotify==0.9.6
PyJWT==2.7.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyRFC3339==1.1
pyserial==3.5
pysmi==0.3.4
pysnmp==4.4.12
pystache==0.6.5
python-dateutil==2.9.0
python-dotenv==0.21.0
python-magic==0.4.27
pytz==2024.1
PyYAML>=6.0
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
s3transfer==0.10.1
scipy==1.13.1
service-identity==24.1.0
setuptools==74.1.2
six==1.16.0
soupsieve==2.6
sqlalchemy>=1.4
ssh-import-id==5.11
sympy==1.13.1
texttable==1.7.0
Twisted==24.7.0
typeguard==4.3.0
typing_extensions==4.12.2
ufoLib2==0.16.0
ujson==5.10.0
urllib3>=1.25.4,<1.27
uvicorn==0.30.3
uvloop==0.19.0
webencodings==0.5.1
websocket-client>=0.32.0,<1.0
wheel==0.44.0
wsproto==1.2.0
zipp==3.20.0